Code/UI/mqtt-listener/requirements.txt
paho-mqtt==1.6.1
asyncpg==0.29.0
